refactor: 서비스 인터페이스 분리 #136
